feat(web): 👻 Player Account | Game Picker
- The user can now audit the game account using the explorer link directly from the stats button.
- Query slightly modified to include the pda.
- GameCard component was refactored to separate the sections for easier maintainability.
