+++ web/app/global.css
@@ -4,5 +4,8 @@
 
 html,
 body {
-  height: 100%;
+  height: 100dvh;
+  overflow: hidden;
+  margin: 0;
+  padding: 0;
 }
